Updated index.html and styles.css
I continued with the stylizing, by changing the h2 element to a <p> and adding some space in between the letters. I also changed the font weight of the h1 element to 800 instead of 700, gave the <p> element some room for the line spacing, and set the whole project up in media queries. It looks good on the desktop simulated view of a mobile device, but I'll test it out.